break app class into different files
